test: MenuControllerTest customGiven 적용
